client/requirements.txt
apiview==1.3.25
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
crypto==1.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
Naked==0.1.31
odfpy==1.4.1
openpyxl==3.0.10
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.10
xlrd==2.0.1
xlwt==1.3.0
